requirement.txt
Flask
qrcode[pil]
pillow
pytest